DAOS-7174 doc: add label support to admin and user guide
- Change command lines in both admin and user guide to
  use labels for both pools and containers instead of UUID.
- Add workflow section to user documentation
- Rename POSIX to filesystem and provide more details in the
  dfuse section.
- Remove references to `daos_server network scan` in favor of
  `dmg network scan`
- Several doc cleanups.

A few dmg and daos utility interface changes:
- support 'daos cont ls/list' to list containers
- support lsattr/delattr/setattr for attributes
- implement other useful shortcuts
