+++ src/control/cmd/daos/pool.go
@@ -8,7 +8,6 @@ package main
 
 import (
 	"fmt"
-	"io"
 	"os"
 	"strings"
 	"unsafe"
@@ -21,7 +20,6 @@ import (
 	"github.com/daos-stack/daos/src/control/common/proto/convert"
 	mgmtpb "github.com/daos-stack/daos/src/control/common/proto/mgmt"
 	"github.com/daos-stack/daos/src/control/lib/control"
-	"github.com/daos-stack/daos/src/control/lib/txtfmt"
 	"github.com/daos-stack/daos/src/control/lib/ui"
 )
 
@@ -139,110 +137,12 @@ func (cmd *poolBaseCmd) getAttr(name string) (*attribute, error) {
 }
 
 type poolCmd struct {
-	ListContainers poolContainersListCmd `command:"list-containers" alias:"list-cont" alias:"ls" description:"list all containers in pool"`
 	Query          poolQueryCmd          `command:"query" description:"query pool info"`
-	ListAttrs      poolListAttrsCmd      `command:"list-attributes" alias:"list-attrs" description:"list pool user-defined attributes"`
-	GetAttr        poolGetAttrCmd        `command:"get-attribute" alias:"get-attr" description:"get pool user-defined attribute"`
-	SetAttr        poolSetAttrCmd        `command:"set-attribute" alias:"set-attr" description:"set pool user-defined attribute"`
-	DelAttr        poolDelAttrCmd        `command:"delete-attribute" alias:"del-attr" description:"delete pool user-defined attribute"`
+	ListAttrs      poolListAttrsCmd      `command:"list-attr" alias:"list-attrs" alias:"lsattr" description:"list pool user-defined attributes"`
+	GetAttr        poolGetAttrCmd        `command:"get-attr" alias:"getattr" description:"get pool user-defined attribute"`
+	SetAttr        poolSetAttrCmd        `command:"set-attr" alias:"setattr" description:"set pool user-defined attribute"`
+	DelAttr        poolDelAttrCmd        `command:"del-attr" alias:"delattr" description:"delete pool user-defined attribute"`
 	AutoTest       poolAutoTestCmd       `command:"autotest" description:"verify setup with smoke tests"`
-}
-
-type poolContainersListCmd struct {
-	poolBaseCmd
-}
-
-func poolListContainers(hdl C.daos_handle_t) ([]*ContainerID, error) {
-	extra_cont_margin := C.size_t(16)
-
-	// First call gets the current number of containers.
-	var ncont C.daos_size_t
-	rc := C.daos_pool_list_cont(hdl, &ncont, nil, nil)
-	if err := daosError(rc); err != nil {
-		return nil, errors.Wrap(err, "pool list containers failed")
-	}
-
-	// No containers.
-	if ncont == 0 {
-		return nil, nil
-	}
-
-	var cConts *C.struct_daos_pool_cont_info
-	// Extend ncont with a safety margin to account for containers
-	// that might have been created since the first API call.
-	ncont += extra_cont_margin
-	cConts = (*C.struct_daos_pool_cont_info)(C.calloc(C.sizeof_struct_daos_pool_cont_info, ncont))
-	if cConts == nil {
-		return nil, errors.New("calloc() for containers failed")
-	}
-	dpciSlice := (*[1 << 30]C.struct_daos_pool_cont_info)(
-		unsafe.Pointer(cConts))[:ncont:ncont]
-	cleanup := func() {
-		C.free(unsafe.Pointer(cConts))
-	}
-
-	rc = C.daos_pool_list_cont(hdl, &ncont, cConts, nil)
-	if err := daosError(rc); err != nil {
-		cleanup()
-		return nil, err
-	}
-
-	out := make([]*ContainerID, ncont)
-	for i := range out {
-		out[i] = new(ContainerID)
-		out[i].UUID = uuid.Must(uuidFromC(dpciSlice[i].pci_uuid))
-		out[i].Label = C.GoString(&dpciSlice[i].pci_label[0])
-	}
-
-	return out, nil
-}
-
-func printContainerList(out io.Writer, contIDs []*ContainerID) {
-	if len(contIDs) == 0 {
-		fmt.Fprintf(out, "No containers.\n")
-		return
-	}
-
-	uuidTitle := "UUID"
-	labelTitle := "Label"
-	titles := []string{uuidTitle, labelTitle}
-
-	table := []txtfmt.TableRow{}
-	for _, id := range contIDs {
-		table = append(table,
-			txtfmt.TableRow{
-				uuidTitle:  id.UUID.String(),
-				labelTitle: id.Label,
-			})
-	}
-
-	tf := txtfmt.NewTableFormatter(titles...)
-	tf.InitWriter(out)
-	tf.Format(table)
-}
-
-func (cmd *poolContainersListCmd) Execute(_ []string) error {
-	cleanup, err := cmd.resolveAndConnect(nil)
-	if err != nil {
-		return err
-	}
-	defer cleanup()
-
-	contIDs, err := poolListContainers(cmd.cPoolHandle)
-	if err != nil {
-		return errors.Wrapf(err,
-			"unable to list containers for pool %s", cmd.PoolID())
-	}
-
-	if cmd.jsonOutputEnabled() {
-		return cmd.outputJSON(contIDs, nil)
-	}
-
-	var bld strings.Builder
-	printContainerList(&bld, contIDs)
-	cmd.log.Info(bld.String())
-
-	return nil
 }
 
 type poolQueryCmd struct {
